Implements the qint encoding in Rust and adds unit tests.
The qint encoding scheme is a variable-length encoding scheme for integers. It uses a leading byte to
indicate the number of bytes used to represent the following integers. Based on that leading byte, up to 4 integers
can be encoded in a single byte buffer. The leading byte is the first byte of the buffer. It encodes each
variable integer length in a 2-bit field. The first 2 bits represent the first integer. The next 2 bits
represent the second integer, and so on.

We don't bind the c-side in this PRS as the qint is encode/decode is only needed in the inverted_index encoder / decoder definitions.